Improve short circuit conditions, remove debug prints.
Best run yet! Beats 100% with time, beats 77.75% with space. Since this is Python, I think having a space difference of 17.2mb for the best to 17.7mb for my answer is acceptable. Write it in C if you care more. (Also yes, leetcode uses mb for their size value which is strange, why not MB?)